split rustyline helper to a separate file
